+++ src/components/ui/Empty/Empty.tsx
@@ -30,7 +30,12 @@ export const Empty = <TRow extends BaseEntityDto>(
     <div className="empty-container">
       {iconProps && <FontAwesomeIcon {...iconProps} />}
       <p {...messageProps}>{message}</p>
-      {action && <Action showTooltips={false} showText {...action} />}
+      <div className="flex gap-2 flex-wrap items-center justify-center">
+        {action &&
+          (Array.isArray(action) ? action : [action]).map((a) => (
+            <Action key={a.id} showTooltips={false} showText {...a} />
+          ))}
+      </div>
     </div>
   );
 };